+++ src/Maybe.php
@@ -202,4 +202,17 @@ final class Maybe
     {
         return $this->maybe->toSequence();
     }
+
+    /**
+     * @template V
+     *
+     * @param callable(T): self<V> $just
+     * @param callable(): self<V> $nothing
+     *
+     * @return self<V>
+     */
+    public function eitherWay(callable $just, callable $nothing): self
+    {
+        return $this->maybe->eitherWay($just, $nothing);
+    }
 }